requirements.txt
h5py
tabulate
tqdm
mxnet
yacs
colorlog
cvbase
keras
lmdb
redis
future-fstrings
# tensorflow-gpu
# torch
# mxnet-cu90
# faiss-cpu
# conda create -n py37 python=3.7 pytorch=1.1 anaconda tensorflow-gpu torchvision tensorboard scikit-learn scikit-image opencv bcolz
# conda install python=3.7 anaconda pytorch=1.1 faiss-cpu tensorflow-gpu torchvision tensorboard scikit-learn scikit-image opencv bcolz
gpustat
easydict
msgpack
msgpack-numpy
ipython
matplotlib
pandas
numpy
Pillow
scipy
trash-cli
tensorboardX
thop

